requirements.txt
streamlit
pdfplumber
scikit-learn
streamlit>=1.20
pdfplumber>=0.7.6
scikit-learn>=1.0
numpy
sentence-transformers>=2.2.2; python_version>="3.8"  # optional (install only if you want semantic embeddings)
streamlit
PyPDF2
pillow
